requirements.txt
Django
djangorestframework
python-dotenv
Requests